% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_custom.R
\name{add_filler_samples}
\alias{add_filler_samples}
\title{Create barplot filler samples}
\usage{
add_filler_samples(df, bars_per_facet)
}
\arguments{
\item{df}{A data frame containing the data to be plotted.}

\item{bars_per_facet}{An integer specifying the number of bars per facet.}
}
\value{
A list with:
\describe{
\item{df:}{A data frame including the original and any filler samples.}
\item{filler_names:}{A named character vector with blank labels for filler samples.}
}
}
\description{
Adds filler rows to a data frame to ensure consistent bar counts per facet in a barplot.
}
