% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_reports.R
\name{bar_chart_pos_neg}
\alias{bar_chart_pos_neg}
\title{barplot in reactable}
\usage{
bar_chart_pos_neg(
  label,
  value,
  max_value = 1,
  height = "1rem",
  pos_fill = t_col("tomato", 0.8),
  neg_fill = t_col("yellowgreen", 0.8)
)
}
\arguments{
\item{value}{the height of bar}

\item{max_value}{the bar color}

\item{height}{the alignment of bar}

\item{pos_fill}{the font color}

\item{neg_fill}{the font weight}

\item{lable}{the length of bar}
}
\description{
barplot in reactable
}
