% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_reports.R
\name{bar_style}
\alias{bar_style}
\title{barplot in reactable}
\usage{
bar_style(
  length = 1,
  height = "80\%",
  fill = "#00FFFF7F",
  align = c("right", "left"),
  color = "black",
  fweight = "plain"
)
}
\arguments{
\item{length}{the length of bar}

\item{height}{the height of bar}

\item{fill}{the bar color}

\item{align}{the alignment of bar}

\item{color}{the font color}

\item{fweight}{the font weight}
}
\description{
barplot in reactable
}
