% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded.R
\name{capital_names}
\alias{capital_names}
\title{capitalise the first character in the string}
\usage{
capital_names(x)
}
\arguments{
\item{x}{a vector of strings}
}
\value{
capitalised strings
}
\description{
creating the capitalised names for data frame
}
