% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_init.R
\name{create_config}
\alias{create_config}
\title{create the template of config file}
\usage{
create_config(config_dir)
}
\arguments{
\item{config_dir}{the directory of config file}
}
\description{
create the template of config file
}
