% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_reports.R
\name{create_qc_reports}
\alias{create_qc_reports}
\title{create QC reports}
\usage{
create_qc_reports(
  samplesheet = NULL,
  qc_type = c("plasmid", "screen"),
  qc_dir = NULL
)
}
\arguments{
\item{samplesheet}{the path of sample sheet file}

\item{qc_type}{screen or plasmid}

\item{qc_dir}{the directory of QC plots and outs}
}
\description{
create QC reports
}
