% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_custom.R
\name{facet_wrap_custom}
\alias{facet_wrap_custom}
\title{facet_wrap_custom}
\usage{
facet_wrap_custom(..., scale_overrides = NULL)
}
\arguments{
\item{scale_overrides}{scale_overrides}
}
\description{
facet_wrap_custom
}
