% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_init.R
\name{load_config}
\alias{load_config}
\title{load the user's config file}
\usage{
load_config(config_path)
}
\arguments{
\item{config_path}{the path of config file}
}
\description{
load the user's config file
}
