% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded.R
\name{merge_list_to_dt}
\alias{merge_list_to_dt}
\title{merge a list of data tables into a data table}
\usage{
merge_list_to_dt(list_dt, by_val, join_val)
}
\arguments{
\item{list_dt}{a list of data tables}

\item{by_val}{join data tables by which column}

\item{join_val}{join which column in the data tables}
}
\value{
a data table
}
\description{
merge a list of data tables into a data table
}
