% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outs.R
\name{qcout_expqc_all,experimentQC-method}
\alias{qcout_expqc_all,experimentQC-method}
\title{create all the output files}
\usage{
\S4method{qcout_expqc_all}{experimentQC}(object, out_dir = NULL)
}
\arguments{
\item{object}{experimentQC object}

\item{out_dir}{the output directory}
}
\description{
create all the output files
}
