% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outs.R
\name{qcout_expqc_all}
\alias{qcout_expqc_all}
\title{initialize function}
\usage{
qcout_expqc_all(object, ...)

\S4method{qcout_expqc_all}{experimentQC}(object, out_dir = NULL)
}
\arguments{
\item{object}{experimentQC object}

\item{out_dir}{the output directory}
}
\description{
initialize function

create all the output files
}
