% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outs.R
\name{qcout_expqc_corr,experimentQC-method}
\alias{qcout_expqc_corr,experimentQC-method}
\title{create output file of clustering and correlation results}
\usage{
\S4method{qcout_expqc_corr}{experimentQC}(object, out_dir = NULL)
}
\arguments{
\item{object}{experimentQC object}

\item{out_dir}{the output directory}
}
\description{
create output file of clustering and correlation results
}
