% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outs.R
\name{qcout_expqc_corr}
\alias{qcout_expqc_corr}
\title{initialize function}
\usage{
qcout_expqc_corr(object, ...)

\S4method{qcout_expqc_corr}{experimentQC}(object, out_dir = NULL)
}
\arguments{
\item{object}{experimentQC object}

\item{out_dir}{the output directory}
}
\description{
initialize function

create output file of clustering and correlation results
}
