% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outs.R
\name{qcout_expqc_deseq}
\alias{qcout_expqc_deseq}
\title{initialize function}
\usage{
qcout_expqc_deseq(object, ...)

\S4method{qcout_expqc_deseq}{experimentQC}(object, eqc_type = c("lib", "all"), out_dir = NULL)
}
\arguments{
\item{object}{experimentQC object}

\item{eqc_type}{library counts or all counts}

\item{out_dir}{the output directory}
}
\description{
initialize function

create all the output files
}
