% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outs.R
\name{qcout_samqc_accepted}
\alias{qcout_samqc_accepted}
\title{initialize function}
\usage{
qcout_samqc_accepted(object, ...)

\S4method{qcout_samqc_accepted}{sampleQC}(object, qc_type = c("plasmid", "screen"), out_dir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{qc_type}{screen or plasmid}

\item{out_dir}{the output directory}
}
\description{
initialize function

create output file of library reads stats
}
