% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outs.R
\name{qcout_samqc_all,sampleQC-method}
\alias{qcout_samqc_all,sampleQC-method}
\title{create all the output files}
\usage{
\S4method{qcout_samqc_all}{sampleQC}(object, qc_type = c("plasmid", "screen"), out_dir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{qc_type}{qc type}

\item{out_dir}{the output directory}
}
\description{
create all the output files
}
