% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outs.R
\name{qcout_samqc_pos_anno}
\alias{qcout_samqc_pos_anno}
\title{initialize function}
\usage{
qcout_samqc_pos_anno(object, ...)

\S4method{qcout_samqc_pos_anno}{sampleQC}(object, out_dir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{out_dir}{the output directory}
}
\description{
initialize function

create output file of lof percentages
}
