% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outs.R
\name{qcout_samqc_total,sampleQC-method}
\alias{qcout_samqc_total,sampleQC-method}
\title{create output file of total reads stats}
\usage{
\S4method{qcout_samqc_total}{sampleQC}(object, out_dir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{out_dir}{the output directory}
}
\description{
create output file of total reads stats
}
