% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_expqc_all}
\alias{qcplot_expqc_all}
\title{initialize function}
\usage{
qcplot_expqc_all(object, ...)

\S4method{qcplot_expqc_all}{experimentQC}(object, plot_dir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{plot_dir}{the output plot directory}
}
\description{
initialize function

create all the plot figures
}
