% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_expqc_sample_corr}
\alias{qcplot_expqc_sample_corr}
\title{initialize function}
\usage{
qcplot_expqc_sample_corr(object, ...)

\S4method{qcplot_expqc_sample_corr}{experimentQC}(object, plot_dir = NULL)
}
\arguments{
\item{object}{experimentQC object}

\item{plot_dir}{the output plot directory}
}
\description{
initialize function

create the heatmap of samples
}
