% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_samqc_clusters,sampleQC-method}
\alias{qcplot_samqc_clusters,sampleQC-method}
\title{create the sequence counts and clusters plot}
\usage{
\S4method{qcplot_samqc_clusters}{sampleQC}(
  object,
  qc_type = c("plasmid", "screen"),
  plot_dir = NULL
)
}
\arguments{
\item{object}{sampleQC object}

\item{qc_type}{qc type for plot}

\item{plot_dir}{the output plot directory}
}
\description{
create the sequence counts and clusters plot
}
