% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_samqc_gini}
\alias{qcplot_samqc_gini}
\title{initialize function}
\usage{
qcplot_samqc_gini(object, ...)

\S4method{qcplot_samqc_gini}{sampleQC}(object, plot_dir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{plot_dir}{the output plot directory}
}
\description{
initialize function

create the gini plot
}
