% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_samqc_pos_cov}
\alias{qcplot_samqc_pos_cov}
\title{initialize function}
\usage{
qcplot_samqc_pos_cov(object, ...)

\S4method{qcplot_samqc_pos_cov}{sampleQC}(object, qc_type = c("plasmid", "screen"), plot_dir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{qc_type}{plot type, screen or plasmid}

\item{plot_dir}{the output plot directory}
}
\description{
initialize function

create the position plot
}
