% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_samqc_readlens}
\alias{qcplot_samqc_readlens}
\title{initialize function}
\usage{
qcplot_samqc_readlens(object, ...)

\S4method{qcplot_samqc_readlens}{sampleQC}(object, len_bins = seq(0, 300, 50), plot_dir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{len_bins}{the bins of length distribution}

\item{plot_dir}{the output plot directory}
}
\description{
initialize function

create the read length plot
}
