% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_samqc_total}
\alias{qcplot_samqc_total}
\title{initialize function}
\usage{
qcplot_samqc_total(object, ...)

\S4method{qcplot_samqc_total}{sampleQC}(object, qc_type = c("plasmid", "screen"), plot_dir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{qc_type}{qc type for plot}

\item{plot_dir}{the output plot directory}
}
\value{
a ggplot object
}
\description{
initialize function

create the stats plot
}
