% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentQC.R
\name{run_experiment_qc_lib_lfc,experimentQC-method}
\alias{run_experiment_qc_lib_lfc,experimentQC-method}
\title{run DESeq2 for the list of samples}
\usage{
\S4method{run_experiment_qc_lib_lfc}{experimentQC}(object, pcut, dcut, ecut, ntop)
}
\arguments{
\item{object}{experimentQC object}

\item{pcut}{the padj cutoff}

\item{dcut}{the depleted log2 fold change cutoff}

\item{ecut}{the enriched log2 fold change cutoff}

\item{ntop}{the number of top variances}
}
\value{
object
}
\description{
run DESeq2 for the list of samples
}
