% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleQC.R
\name{run_sample_qc}
\alias{run_sample_qc}
\title{initialize function}
\usage{
run_sample_qc(object, ...)

\S4method{run_sample_qc}{sampleQC}(
  object,
  qc_type = c("plasmid", "screen"),
  cutoff_total = maveqc_config$sqc_total,
  cutoff_missing_per = maveqc_config$sqc_missing,
  cutoff_low_count = maveqc_config$sqc_low_count,
  cutoff_low_sample_per = maveqc_config$sqc_low_sample_per,
  cutoff_accepted = maveqc_config$sqc_accepted,
  cutoff_mapping_per = maveqc_config$sqc_mapping_per,
  cutoff_ref_per = maveqc_config$sqc_ref_per,
  screen_cutoff_library_per = maveqc_config$screen_sqc_library_per,
  plasmid_cutoff_library_per = maveqc_config$plasmid_sqc_library_per,
  cutoff_library_cov = maveqc_config$sqc_library_cov,
  cutoff_low_per = maveqc_config$sqc_low_per,
  cutoff_low_lib_per = maveqc_config$sqc_low_lib_per
)
}
\arguments{
\item{object}{sampleQC object}

\item{qc_type}{plasmid or screen}

\item{cutoff_total}{qc cutoff of the total reads}

\item{cutoff_missing_per}{qc cutoff of the missing variant percentage}

\item{cutoff_low_count}{count cutoff of library reads}

\item{cutoff_low_sample_per}{sample percentage cutoff of library reads}

\item{cutoff_accepted}{qc cutoff of the total accepted reads}

\item{cutoff_mapping_per}{qc cutoff of mapping percentage (ref + pam + library)}

\item{cutoff_ref_per}{qc cutoff of reference percentage}

\item{screen_cutoff_library_per}{qc cutoff of library reads percentage for screen runs}

\item{plasmid_cutoff_library_per}{qc cutoff of library reads percentage for plasmid runs}

\item{cutoff_library_cov}{qc cutoff of library coverage}

\item{cutoff_low_per}{qc cutoff of low abundance percentage for LOF plot}

\item{cutoff_low_lib_per}{qc cutoff of the percentage of library sequences with low abundance for LOF plot}
}
\value{
object
}
\description{
initialize function

run sample QC for the list of samples
}
