% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_custom.R
\name{scale_override}
\alias{scale_override}
\title{scale_override for facet_custom}
\usage{
scale_override(which, scale)
}
\arguments{
\item{which}{which grid in the facet}

\item{scale}{x or y scale}
}
\value{
a structure
}
\description{
scale_override for facet_custom
}
