% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{sge_qc_stats}
\alias{sge_qc_stats}
\title{initialize function}
\usage{
sge_qc_stats(object, ...)

\S4method{sge_qc_stats}{SGE}(object)
}
\arguments{
\item{object}{SGE object}

\item{lowcut}{cutoff which determines the oligo count is low, user's definition}
}
\value{
object
}
\description{
initialize function

format library dependent and independent counts with extra info and remove duplicate/useless info
}
