% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{show_stats}
\alias{show_stats}
\title{initialize function}
\usage{
show_stats(object, ...)

\S4method{show_stats}{SGE}(object)
}
\arguments{
\item{object}{SGE object}
}
\description{
initialize function

show basic stats of a SGE object
}
