% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded.R, R/facet_custom.R
\name{t_col}
\alias{t_col}
\alias{scale_override}
\title{transparent color function}
\usage{
t_col(col, rate)

scale_override(which, scale)
}
\arguments{
\item{col}{color name}

\item{rate}{alpha rate}

\item{which}{which grid in the facet}

\item{scale}{x or y scale}
}
\value{
transparent color

a structure
}
\description{
transparent color function

scale_override for facet_custom
}
